Restore old options class name and command package
OptionsPlugin uses Java Preferences system for persisting options.
This system uses fully qualified class names as keys for data persistence.
