+++ src/main/java/org/ilastik/ilastik4ij/ui/ObjectClassificationCommand.java
@@ -1,4 +1,4 @@
-package org.ilastik.ilastik4ij.cmd;
+package org.ilastik.ilastik4ij.ui;
 
 import net.imagej.Dataset;
 import net.imglib2.type.NativeType;
@@ -7,11 +7,10 @@ import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
-@Plugin(type = Command.class, headless = true, menuPath = "Plugins>ilastik>Run Tracking")
-public final class TrackingCommand<T extends NativeType<T>> extends CommandBase<T> {
+@Plugin(type = Command.class, headless = true, menuPath = "Plugins>ilastik>Run Object Classification Prediction")
+public final class ObjectClassificationCommand<T extends NativeType<T>> extends CommandBase<T> {
     @Parameter(label = "Pixel Probability or Segmentation image")
     public Dataset inputProbOrSegImage;
 
@@ -19,17 +18,12 @@ public final class TrackingCommand<T extends NativeType<T>> extends CommandBase<
     public String secondInputType = ROLE_PROBABILITIES;
 
     @Override
-    protected List<String> workflowArgs() {
-        return Collections.singletonList("--export_source=Tracking-Result");
-    }
-
-    @Override
     protected Map<String, Dataset> workflowInputs() {
         if (ROLE_PROBABILITIES.equals(secondInputType)) {
             return Collections.singletonMap("prediction_maps", inputProbOrSegImage);
         }
         if (ROLE_SEGMENTATION.equals(secondInputType)) {
-            return Collections.singletonMap("binary_image", inputProbOrSegImage);
+            return Collections.singletonMap("segmentation_image", inputProbOrSegImage);
         }
         throw new IllegalStateException("Unexpected value: " + secondInputType);
     }
